requirements.txt
future~=0.18.2
pywin32~=304
requests~=2.28.1
lxml~=4.9.1
python-dotenv~=0.20.0